:recycle: Split account service into directory module by use case
Pure mechanical move of application/src/service/account.rs into a directory module; no behavior change.
