--- application/src/service/account.rs
@@ -1,568 +1,0 @@
-use crate::permission::{
-    account_deactivate, account_edit, account_view, check_permission, instance_moderate,
-};
-use crate::signing_key::CreateSigningKeyUseCase;
-use crate::transfer::account::{AccountDto, CreateAccountDto, UpdateAccountDto};
-use crate::transfer::pagination::{apply_pagination, Pagination};
-use adapter::crypto::DependOnSigningKeyGenerator;
-use adapter::processor::account::{
-    AccountCommandProcessor, AccountQueryProcessor, CreateAccountParam,
-    DependOnAccountCommandProcessor, DependOnAccountQueryProcessor, UpdateAccountParam,
-};
-use adapter::processor::profile::{
-    CreateProfileParam, DependOnProfileCommandProcessor, ProfileCommandProcessor,
-};
-use error_stack::Report;
-use kernel::interfaces::config::DependOnPublicBaseUrl;
-use kernel::interfaces::crypto::{DependOnPasswordProvider, SigningAlgorithm};
-use kernel::interfaces::database::DatabaseConnection;
-use kernel::interfaces::event::EventApplier;
-use kernel::interfaces::event_store::{AccountEventStore, DependOnAccountEventStore};
-use kernel::interfaces::permission::{
-    AccountRelation, DependOnPermissionChecker, DependOnPermissionWriter, PermissionWriter,
-    RelationTarget,
-};
-use kernel::interfaces::repository::DependOnSigningKeyRepository;
-use kernel::prelude::entity::{
-    Account, AccountId, AccountIsBot, AccountName, AuthAccountId, EventId, EventVersion,
-    ModerationReason, Nanoid, Profile, ProfileDisplayName,
-};
-use kernel::KernelError;
-use std::future::Future;
-
-pub trait GetAccountUseCase:
-    'static + Sync + Send + DependOnAccountQueryProcessor + DependOnPermissionChecker
-{
-    // find_by_auth_id returns only accounts owned by the authenticated user,
-    // so no additional permission check is needed.
-    fn get_all_accounts(
-        &self,
-        auth_account_id: &AuthAccountId,
-        Pagination {
-            direction,
-            cursor,
-            limit,
-        }: Pagination<String>,
-    ) -> impl Future<Output = error_stack::Result<Option<Vec<AccountDto>>, KernelError>> + Send
-    {
-        async move {
-            let mut transaction = self.database_connection().get_executor().await?;
-            let accounts = self
-                .account_query_processor()
-                .find_by_auth_id(&mut transaction, auth_account_id)
-                .await?;
-            let cursor = if let Some(cursor) = cursor {
-                let id: Nanoid<Account> = Nanoid::new(cursor);
-                self.account_query_processor()
-                    .find_by_nanoid(&mut transaction, &id)
-                    .await?
-            } else {
-                None
-            };
-            let accounts = apply_pagination(accounts, limit, cursor, direction);
-            Ok(Some(accounts.into_iter().map(AccountDto::from).collect()))
-        }
-    }
-
-    fn get_accounts_by_ids(
-        &self,
-        auth_account_id: &AuthAccountId,
-        ids: Vec<String>,
-    ) -> impl Future<Output = error_stack::Result<Vec<AccountDto>, KernelError>> + Send {
-        async move {
-            let mut transaction = self.database_connection().get_executor().await?;
-
-            let nanoids: Vec<Nanoid<Account>> =
-                ids.into_iter().map(Nanoid::<Account>::new).collect();
-            let accounts = self
-                .account_query_processor()
-                .find_by_nanoids(&mut transaction, &nanoids)
-                .await?;
-
-            let mut result = Vec::new();
-            for account in accounts {
-                if check_permission(self, auth_account_id, &account_view(account.id()))
-                    .await
-                    .is_ok()
-                {
-                    result.push(AccountDto::from(account));
-                }
-            }
-
-            Ok(result)
-        }
-    }
-}
-
-impl<T> GetAccountUseCase for T where
-    T: 'static + DependOnAccountQueryProcessor + DependOnPermissionChecker
-{
-}
-
-pub trait CreateAccountUseCase:
-    'static
-    + Sync
-    + Send
-    + DependOnAccountCommandProcessor
-    + DependOnProfileCommandProcessor
-    + DependOnPasswordProvider
-    + DependOnSigningKeyGenerator
-    + DependOnPermissionWriter
-    + DependOnSigningKeyRepository
-    + DependOnPublicBaseUrl
-{
-    fn create_account(
-        &self,
-        auth_account_id: AuthAccountId,
-        dto: CreateAccountDto,
-    ) -> impl Future<Output = error_stack::Result<AccountDto, KernelError>> + Send {
-        async move {
-            let mut transaction = self.database_connection().get_executor().await?;
-
-            let account_name = AccountName::new(dto.name);
-            let account_is_bot = AccountIsBot::new(dto.is_bot);
-
-            let display_name = ProfileDisplayName::new(account_name.as_ref().to_string());
-
-            let account = self
-                .account_command_processor()
-                .create(
-                    &mut transaction,
-                    CreateAccountParam {
-                        name: account_name,
-                        is_bot: account_is_bot,
-                        auth_account_id: auth_account_id.clone(),
-                    },
-                )
-                .await?;
-
-            self.profile_command_processor()
-                .create(
-                    &mut transaction,
-                    CreateProfileParam {
-                        account_id: account.id().clone(),
-                        display_name: Some(display_name),
-                        summary: None,
-                        icon: None,
-                        banner: None,
-                        nano_id: Nanoid::<Profile>::default(),
-                    },
-                )
-                .await?;
-
-            self.permission_writer()
-                .create_relation(
-                    &RelationTarget::Account {
-                        account_id: account.id().clone(),
-                        relation: AccountRelation::Owner,
-                    },
-                    &auth_account_id,
-                )
-                .await?;
-
-            self.create(
-                account.id().clone(),
-                account.nanoid(),
-                SigningAlgorithm::Rsa2048,
-            )
-            .await?;
-
-            Ok(AccountDto::from(account))
-        }
-    }
-}
-
-impl<T> CreateAccountUseCase for T where
-    T: 'static
-        + DependOnAccountCommandProcessor
-        + DependOnProfileCommandProcessor
-        + DependOnPasswordProvider
-        + DependOnSigningKeyGenerator
-        + DependOnPermissionWriter
-        + DependOnSigningKeyRepository
-        + DependOnPublicBaseUrl
-{
-}
-
-pub trait UpdateAccountUseCase:
-    'static
-    + Sync
-    + Send
-    + DependOnAccountCommandProcessor
-    + DependOnAccountQueryProcessor
-    + DependOnAccountEventStore
-    + DependOnPermissionChecker
-{
-    fn update_account(
-        &self,
-        auth_account_id: &AuthAccountId,
-        dto: UpdateAccountDto,
-    ) -> impl Future<Output = error_stack::Result<(), KernelError>> + Send {
-        async move {
-            let mut transaction = self.database_connection().get_executor().await?;
-
-            let nanoid = Nanoid::<Account>::new(dto.account_nanoid);
-            let projection = self
-                .account_query_processor()
-                .find_by_nanoid_unfiltered(&mut transaction, &nanoid)
-                .await?
-                .ok_or_else(|| {
-                    Report::new(KernelError::NotFound).attach_printable(format!(
-                        "Account not found with nanoid: {}",
-                        nanoid.as_ref()
-                    ))
-                })?;
-
-            check_permission(self, auth_account_id, &account_edit(projection.id())).await?;
-
-            let account_id = projection.id().clone();
-            let (account, current_version) =
-                rehydrate_account(self, &mut transaction, &account_id).await?;
-
-            let is_bot = match dto.is_bot {
-                kernel::prelude::entity::FieldAction::Unchanged => *account.is_bot().as_ref(),
-                kernel::prelude::entity::FieldAction::Clear => false,
-                kernel::prelude::entity::FieldAction::Set(value) => value,
-            };
-
-            if !account.status().is_active() {
-                return Err(Report::new(KernelError::Rejected)
-                    .attach_printable("Cannot modify a suspended or banned account"));
-            }
-            if account.deleted_at().is_some() {
-                return Err(
-                    Report::new(KernelError::Rejected).attach_printable("Account is deactivated")
-                );
-            }
-
-            self.account_command_processor()
-                .update(
-                    &mut transaction,
-                    UpdateAccountParam {
-                        account_id,
-                        is_bot: AccountIsBot::new(is_bot),
-                        current_version,
-                    },
-                )
-                .await?;
-
-            Ok(())
-        }
-    }
-}
-
-impl<T> UpdateAccountUseCase for T where
-    T: 'static
-        + DependOnAccountCommandProcessor
-        + DependOnAccountQueryProcessor
-        + DependOnAccountEventStore
-        + DependOnPermissionChecker
-{
-}
-
-pub trait DeactivateAccountUseCase:
-    'static
-    + Sync
-    + Send
-    + DependOnAccountCommandProcessor
-    + DependOnAccountQueryProcessor
-    + DependOnAccountEventStore
-    + DependOnPermissionChecker
-    + DependOnPermissionWriter
-{
-    fn deactivate_account(
-        &self,
-        auth_account_id: &AuthAccountId,
-        account_id: String,
-    ) -> impl Future<Output = error_stack::Result<(), KernelError>> + Send {
-        async move {
-            let mut transaction = self.database_connection().get_executor().await?;
-
-            let nanoid = Nanoid::<Account>::new(account_id);
-            let projection = self
-                .account_query_processor()
-                .find_by_nanoid(&mut transaction, &nanoid)
-                .await?
-                .ok_or_else(|| {
-                    Report::new(KernelError::NotFound).attach_printable(format!(
-                        "Account not found with nanoid: {}",
-                        nanoid.as_ref()
-                    ))
-                })?;
-
-            check_permission(self, auth_account_id, &account_deactivate(projection.id())).await?;
-
-            let account_id = projection.id().clone();
-            let (_account, current_version) =
-                rehydrate_account(self, &mut transaction, &account_id).await?;
-            self.account_command_processor()
-                .deactivate(&mut transaction, account_id.clone(), current_version)
-                .await?;
-
-            for relation in [
-                AccountRelation::Owner,
-                AccountRelation::Editor,
-                AccountRelation::Signer,
-            ] {
-                self.permission_writer()
-                    .delete_relation(
-                        &RelationTarget::Account {
-                            account_id: account_id.clone(),
-                            relation,
-                        },
-                        auth_account_id,
-                    )
-                    .await?;
-            }
-
-            Ok(())
-        }
-    }
-}
-
-impl<T> DeactivateAccountUseCase for T where
-    T: 'static
-        + DependOnAccountCommandProcessor
-        + DependOnAccountQueryProcessor
-        + DependOnAccountEventStore
-        + DependOnPermissionChecker
-        + DependOnPermissionWriter
-{
-}
-
-pub trait SuspendAccountUseCase:
-    'static
-    + Sync
-    + Send
-    + DependOnAccountCommandProcessor
-    + DependOnAccountQueryProcessor
-    + DependOnAccountEventStore
-    + DependOnPermissionChecker
-{
-    fn suspend_account(
-        &self,
-        auth_account_id: &AuthAccountId,
-        account_id: String,
-        reason: String,
-        expires_at: Option<time::OffsetDateTime>,
-    ) -> impl Future<Output = error_stack::Result<(), KernelError>> + Send {
-        async move {
-            ModerationReason::new(reason.as_str()).validate()?;
-            let mut transaction = self.database_connection().get_executor().await?;
-
-            let nanoid = Nanoid::<Account>::new(account_id);
-            let projection = self
-                .account_query_processor()
-                .find_by_nanoid_unfiltered(&mut transaction, &nanoid)
-                .await?
-                .ok_or_else(|| {
-                    Report::new(KernelError::NotFound).attach_printable(format!(
-                        "Account not found with nanoid: {}",
-                        nanoid.as_ref()
-                    ))
-                })?;
-
-            check_permission(self, auth_account_id, &instance_moderate()).await?;
-
-            if let Some(exp) = expires_at {
-                if exp <= time::OffsetDateTime::now_utc() {
-                    return Err(Report::new(KernelError::Rejected)
-                        .attach_printable("expires_at must be in the future"));
-                }
-            }
-
-            let account_id = projection.id().clone();
-            let (account, current_version) =
-                rehydrate_account(self, &mut transaction, &account_id).await?;
-
-            if !account.status().is_active() {
-                return Err(
-                    Report::new(KernelError::Rejected).attach_printable("Account is not active")
-                );
-            }
-            if account.deleted_at().is_some() {
-                return Err(
-                    Report::new(KernelError::Rejected).attach_printable("Account is deactivated")
-                );
-            }
-
-            self.account_command_processor()
-                .suspend(
-                    &mut transaction,
-                    account_id,
-                    reason,
-                    expires_at,
-                    current_version,
-                )
-                .await?;
-
-            Ok(())
-        }
-    }
-}
-
-impl<T> SuspendAccountUseCase for T where
-    T: 'static
-        + DependOnAccountCommandProcessor
-        + DependOnAccountQueryProcessor
-        + DependOnAccountEventStore
-        + DependOnPermissionChecker
-{
-}
-
-pub trait UnsuspendAccountUseCase:
-    'static
-    + Sync
-    + Send
-    + DependOnAccountCommandProcessor
-    + DependOnAccountQueryProcessor
-    + DependOnAccountEventStore
-    + DependOnPermissionChecker
-{
-    fn unsuspend_account(
-        &self,
-        auth_account_id: &AuthAccountId,
-        account_id: String,
-    ) -> impl Future<Output = error_stack::Result<(), KernelError>> + Send {
-        async move {
-            let mut transaction = self.database_connection().get_executor().await?;
-
-            let nanoid = Nanoid::<Account>::new(account_id);
-            let projection = self
-                .account_query_processor()
-                .find_by_nanoid_unfiltered(&mut transaction, &nanoid)
-                .await?
-                .ok_or_else(|| {
-                    Report::new(KernelError::NotFound).attach_printable(format!(
-                        "Account not found with nanoid: {}",
-                        nanoid.as_ref()
-                    ))
-                })?;
-
-            check_permission(self, auth_account_id, &instance_moderate()).await?;
-
-            let account_id = projection.id().clone();
-            let (account, current_version) =
-                rehydrate_account(self, &mut transaction, &account_id).await?;
-
-            if !account.status().is_suspended() {
-                return Err(
-                    Report::new(KernelError::Rejected).attach_printable("Account is not suspended")
-                );
-            }
-
-            self.account_command_processor()
-                .unsuspend(&mut transaction, account_id, current_version)
-                .await?;
-
-            Ok(())
-        }
-    }
-}
-
-impl<T> UnsuspendAccountUseCase for T where
-    T: 'static
-        + DependOnAccountCommandProcessor
-        + DependOnAccountQueryProcessor
-        + DependOnAccountEventStore
-        + DependOnPermissionChecker
-{
-}
-
-pub trait BanAccountUseCase:
-    'static
-    + Sync
-    + Send
-    + DependOnAccountCommandProcessor
-    + DependOnAccountQueryProcessor
-    + DependOnAccountEventStore
-    + DependOnPermissionChecker
-{
-    fn ban_account(
-        &self,
-        auth_account_id: &AuthAccountId,
-        account_id: String,
-        reason: String,
-    ) -> impl Future<Output = error_stack::Result<(), KernelError>> + Send {
-        async move {
-            ModerationReason::new(reason.as_str()).validate()?;
-            let mut transaction = self.database_connection().get_executor().await?;
-
-            let nanoid = Nanoid::<Account>::new(account_id);
-            let projection = self
-                .account_query_processor()
-                .find_by_nanoid_unfiltered(&mut transaction, &nanoid)
-                .await?
-                .ok_or_else(|| {
-                    Report::new(KernelError::NotFound).attach_printable(format!(
-                        "Account not found with nanoid: {}",
-                        nanoid.as_ref()
-                    ))
-                })?;
-
-            check_permission(self, auth_account_id, &instance_moderate()).await?;
-
-            let account_id = projection.id().clone();
-            let (account, current_version) =
-                rehydrate_account(self, &mut transaction, &account_id).await?;
-
-            if account.status().is_banned() {
-                return Err(Report::new(KernelError::Rejected)
-                    .attach_printable("Account is already banned"));
-            }
-            if account.deleted_at().is_some() {
-                return Err(
-                    Report::new(KernelError::Rejected).attach_printable("Account is deactivated")
-                );
-            }
-
-            self.account_command_processor()
-                .ban(&mut transaction, account_id, reason, current_version)
-                .await?;
-
-            Ok(())
-        }
-    }
-}
-
-impl<T> BanAccountUseCase for T where
-    T: 'static
-        + DependOnAccountCommandProcessor
-        + DependOnAccountQueryProcessor
-        + DependOnAccountEventStore
-        + DependOnPermissionChecker
-{
-}
-
-pub(crate) async fn rehydrate_account<T>(
-    deps: &T,
-    executor: &mut <<T as kernel::interfaces::database::DependOnDatabaseConnection>::DatabaseConnection as DatabaseConnection>::Executor,
-    account_id: &AccountId,
-) -> error_stack::Result<(Account, EventVersion<Account>), KernelError>
-where
-    T: DependOnAccountEventStore + ?Sized,
-{
-    let event_id = EventId::from(account_id.clone());
-    let events = deps
-        .account_event_store()
-        .find_by_id(executor, &event_id, None)
-        .await?;
-    if events.is_empty() {
-        return Err(Report::new(KernelError::NotFound).attach_printable(format!(
-            "No events found for account: {}",
-            account_id.as_ref()
-        )));
-    }
-    let mut account: Option<Account> = None;
-    for event in events {
-        Account::apply(&mut account, event)?;
-    }
-    let account = account.ok_or_else(|| {
-        Report::new(KernelError::NotFound).attach_printable(format!(
-            "Account aggregate could not be reconstructed for: {}",
-            account_id.as_ref()
-        ))
-    })?;
-    let current_version = account.version().clone();
-    Ok((account, current_version))
-}
